[layout] Fix min/max content sizes with %-padding.
Another thing I just noticed. See test-case. Its possible for the "root"
which is haaving its min/max content size being computed to resolve
%-padding block-start/end. We were doing this, however our cache was
invalid when the width of our containing-block changed.

This bails on the min/max sizes cache if any %-padding is present.

We'll be able to add an additional check here soon to reclaim a little
bit of the cache hits for this case.

Change-Id: I78fb85a2f3e4a347a27ca30c49cc5ee12221c063