[css-grid] Fix that Replaced Layout doesn't display with a property called |max-width|
The replaced which is the grid item doesn't display with max-width
this type of value is the percentage. Because when the grid container
requests the preferred width of a grid item which is replaced such
as image and so on, replaced grid item returns 0 as width on Constrain-
IntrinsicSizeToMinMax. This patch returns the intrinsic height to
compute the width with the aspect ratio.

Bug: 1127715
Change-Id: Ie448dc1b9ea8af62407b02af89d722054bdce00c