[SPCWPT] Getting payment-credential should cause NotAllowedError
The SPC spec specifies that getting credential with payment extension
should cause NotAllowedError[1]:

"When making an assertion with a AuthenticationExtensionsPaymentInputs
extension_inputs: If not in a "secure-payment-confirmation" payment
handler, return a "NotAllowedError" DOMException..."

But the implementation throws NotSupportedError in this case. This CL
aligns the implementation to the spec for the error type.

[1] https://www.w3.org/TR/secure-payment-confirmation/#sctn-payment-extension-registration

Bug: 1249235
Change-Id: I977ca63d254dd4e2e7b346cae423d001c1409d47