Fix content-visibility DCHECK in HTMLMarqueeElement::GetMetrics
This patch DOES NOT fix the crash, I just want to talk about what we
should do first.

This DCHECK is getting hit when GetMetrics accesses computed style:
https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/core/dom/element.cc;l=5156-5157;drc=8b590b20efc5a60f0be016111c242cb8e8693e70

The comments there say that if I use UpdateStyleAndLayoutForNode then I
should be good, but its still crashing... does the adjacent DisplayLocked
node also need to be unlocked somehow? Or does the accessing of style
in GetMetrics need to be updated?

After some more debugging, I see that the DCHECK is getting hit because
one of the nodes inside the marquee's user agent shadowdom has
NeedsStyleRecalc set to true after ComputeComputedStyle is called.
Assuming this is allowed to happen, maybe the DCHECK needs to be updated
to consider this case...?

Fixed: 1251931
Change-Id: I00382a6950f0f019f61f4e4bc43c5339146b6a2f