Prerender: Upstream a basic test for SpeculationRules triggered prerendering to WPT
This CL upstreams a basic prerender test that confirms
`document.prerendering` and `document.onprerenderingchange` event on
pages prerendered by SpeculationRules to the WPT repository.

> Directory structure

This introduces a new directory speculation-rules/ and a sub directory
speculation-rules/prerender/. Tests for other pre* features triggered by
SpeculationRules will be placed in subdirectories of speculation-rules/.

> Feature detection

The test checks if SpeculationRules is available using
`HTMLScriptElement.supports('speculationrules')`. This doesn't exactly
check if SpeculationRules triggered prerendering is available, but
currently this is the only way to (roughly) detect the feature and still
useful for avoiding timeout on other user agents.

Change-Id: I3fbd729de42f560b9157ff1a3b4ba6e4e973a0d6
Bug: 1253158