Create WPT for partitioned service workers
Add a test that verifies that services workers are partitioned and
cannot interact with each other when third-party partitioning is
enabled.

This CL verifies the enabled feature case by running the test in a
virtual test suite.

The test expectations are also updated to expect the test to fail in
the standard suite and pass in the virtual suite.

I manually verified this test by removing the window and iframe
cleanup/removals, running it in content_shell, and verifying that the
windows displayed the appropriate text demonstrating that the SW did
indeed get used.

Bug: 1246549
Change-Id: I22450e3abf496e4e011cc60ef7cfdf04acfdbcda