FSA: Make removal of files compatible with file locking
Removing a file requires acquiring a write lock at the URL.
The lock is released once the operation completes.

remove() requires acquiring an exclusive lock at the URL.

For the sake of backwards compatibility, removeEntry() only requires
acquiring a shared lock. This means that you cannot remove a file with
an open access handle, but you CAN remove a file with open writables.

Removal of directories is not yet compatible with file locking.

Bug: 1114923, 1254078
Change-Id: I1eee08deaff9508e3c90d78ae35b2c181bb18122